pyproject.toml
[build-system]
requires = ["poetry"]
build-backend = "poetry.masonry.api"

[tool]
[tool.isort]
profile = "black"
combine_as_imports = true
multi_line_output = 3
line_length = 120

[tool.poetry]
name = "lunavl"
version = "0.3.0"
description = "Python interface for VisionLabs Luna platform"
authors = ["VisionLabs <m.limonov@visionlabs.ru>"]
repository = "https://github.com/matemax/lunasdk"
[tool.poetry.dependencies]
python = ">=3.8"
numpy = "*"
requests = "*"
Pillow = {version = "*", markers = "sys_platform == 'win32' or sys_platform == 'darwin'"}
pillow-simd = {version = "*", markers = "sys_platform != 'win32' and sys_platform != 'darwin'"}

# [tool.poetry.dependencies.FaceEngine]
# git = "ssh://git@git.visionlabs.ru/face-recognition/FaceEnginePythonBindings.git"
# tag = "v.5.6.0.1"

[tool.poetry.dev-dependencies]
sphinx = ">=2.0.0"
pytest = ">=4.4"
pytest-subtests = "*"
sphinx-autodoc-typehints = "*"
pre-commit = "^1.18"

black = {version = "^20.8b1",allow-prereleases = true}
jsonschema = "^3.2.0"
mypy = "^0.931"
flake8 = "^3.7.9"
bandit = "^1.6.2"


[tool.black]
line-length = 120
target-version = ['py37', 'py36', 'py35']
